feat: cria TurmaService e TurmaController, ajusta tipo dos ids para uso com sqlite

+++ src/main/java/com/project/schoolmanagementapi/dto/aluno/AlunoRequest.java
@@ -3,6 +3,6 @@ package com.project.schoolmanagementapi.dto.aluno;
 public record AlunoRequest(
         String nome,
         String matricula,
-        Long turmaId
+        Integer turmaId
 ) {
 }
