+++ src/main/java/com/project/schoolmanagementapi/repository/DisciplinaRepository.java
@@ -3,5 +3,5 @@ package com.project.schoolmanagementapi.repository;
 import com.project.schoolmanagementapi.model.Disciplina;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisciplinaRepository extends JpaRepository<Disciplina,Long> {
+public interface DisciplinaRepository extends JpaRepository<Disciplina,Integer> {
 }
