+++ src/main/java/com/project/schoolmanagementapi/repository/TurmaRepository.java
@@ -3,5 +3,5 @@ package com.project.schoolmanagementapi.repository;
 import com.project.schoolmanagementapi.model.Turma;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TurmaRepository extends JpaRepository<Turma, Long> {
+public interface TurmaRepository extends JpaRepository<Turma, Integer> {
 }
